big refacotring / splicing data to other files / long distance searching waypoints
