+++ d2bs/kolbot/libs/modules/Town.js
@@ -1956,6 +1956,8 @@
 
 	const act = [{}, {}, {}, {}, {}];
 
+	Town.ActData = act;
+
 	Town.initialize = function () {
 		//print("Initialize town " + me.act);
